+++ test/options.js
@@ -66,9 +66,7 @@ describe('options', function () {
   });
 
   it('#setOptions for global options', function () {
-    this.rt = new RequestTarget({
-      callAllHandlers: false
-    });
+    this.rt = new RequestTarget({ callAllHandlers: false });
 
     const called = [];
 
@@ -85,9 +83,7 @@ describe('options', function () {
   });
 
   it('#setOptions for request options', function () {
-    this.rt = new RequestTarget({
-      callAllHandlers: false
-    });
+    this.rt = new RequestTarget({ callAllHandlers: false });
 
     const called = [];
 
@@ -108,5 +104,41 @@ describe('options', function () {
     this.rt.on('boom', () => void called.push(2));
     this.rt.request('boom');
     expect(called).to.eql([1]);
+  });
+
+  it('options.getAllResults with synchronous results', function () {
+    this.rt = new RequestTarget({ getAllResults: true });
+
+    this.rt.on('boom', () => 1);
+    this.rt.on('boom', () => 3);
+    this.rt.on('boom', () => 5);
+
+    const result = this.rt.request('boom');
+    expect(result).to.eql([1, 3, 5]);
+  })
+
+  it('options.getAllResults with asynchronous results and options.autoPromiseAll', async function () {
+    this.rt = new RequestTarget({ getAllResults: true, autoPromiseAll: true });
+
+    this.rt.on('boom', () => Promise.resolve(1));
+    this.rt.on('boom', () => Promise.resolve(3));
+    this.rt.on('boom', () => Promise.resolve(5));
+
+    const result = await this.rt.request('boom');
+    expect(result).to.eql([1, 3, 5]);
+  });
+
+  it('options.getAllResults with asynchronous results and options.autoPromiseAll as false', function () {
+    this.rt = new RequestTarget({ getAllResults: true, autoPromiseAll: false });
+
+    this.rt.on('boom', () => Promise.resolve(1));
+    this.rt.on('boom', () => Promise.resolve(3));
+    this.rt.on('boom', () => Promise.resolve(5));
+
+    const result = this.rt.request('boom');
+    expect(result).to.be.an('array');
+    expect(result[0]).to.be.a('promise');
+    expect(result[1]).to.be.a('promise');
+    expect(result[2]).to.be.a('promise');
   });
 });