+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.3.0
+
+- Add `getAllResults` option.
+- Add `autoPromiseAll` option.
+
 ## 2.2.0
 
 - Add `callAllHandlers` option.